[ABI Validation] Added dependency on checkLegacyAbi from `check` task
Now, even if ABI validation is enabled in the Gradle project, it is not called automatically when running the check task.

This is counterintuitive, at least because the task is also called checkLegacyAbi.

The check task description also says: Runs all checks. Since checkLegacyAbi is a checking task, we have to run it too.

#KT-80614 Fixed