[IR][K/N] Fixed one problem with caches
If a fake-override function overrides a private real function, calling it from another
module or file could lead to a linkage error. This is because the codegen short-cuts calls
to fake-override functions directly to the actual callees, which is in this example a private
funtion, it has private linkage, causing the linkage error.

This commit fixes this by adding a special accessor for each such fake-override function.

 #KT-52498 Fixed