[Analysis API] Introduce API for finding KDoc for KaSymbol and KtElement
KT-63339

Some Analysis API users (IDE, Dokka) need to access KDoc's of the
symbols. Previously, each user implemented its own wrapper functions on
top of AA. This commit provides a unified API for accessing KDoc.
Implementation is based on the current implementation of the same
functionality in IDE, as it is considered a 'ground truth.'