[kx.serialization] Allow star projections as non-serializable type params for Sealed and Polymorphic serializers.
The bug in K2 checker (see #KT-73107 and previous commit) allowed them to be used in K2 for some time.
While generally it is incorrect, Sealed and PolymorphicSerializer do not need type arguments (except #KT-81747) and it is fine to use star projections there.
See examples in the added test and also in KotlinCompilerArgument.kt

Also https://github.com/Kotlin/kotlinx.serialization/issues/1501