[ABI Validation] Removed the ABI variants functionality
At the moment, during the discussion, it was decided to abandon the concept of dump variants.

There was no widespread request for this feature, but at the same time, the word 'variant' can be confusing, because it is a very common word.

This feature is not required to stabilize DSL and can be added later when real use cases are available.

#KT-80827 Fixed