Reflection: fix javaType for local delegated property accessors
Regarding KT-81899: property setter always returns `Unit`, but types in
kotlin-reflect are position-dependent, so just returning
`typeOf<Unit>()` is not good enough, because its javaType will be the
class object for `kotlin.Unit`. So, we create a special Unit type whose
javaType returns `void`.

Regarding KT-28170/KT-81938: previously, javaType was extracted from
`caller`, but local delegated properties have a special caller that does
not allow calls. In fact, there's no need to use `caller`, because
javaType for the property can be computed directly from the KType (there
are no cases like `Unit` -> `void` here), getter's type is equal to the
property type, and setter's type is again the `Unit` mapped to `void`.

Also, move code that extracts continuation argument from
`DescriptorKCallable` to `DescriptorKFunction`, because only functions
can be suspend. The only suspend property is `coroutineContext`, and
references to it are not allowed.

 #KT-28170 Fixed
 #KT-81899 Fixed
 #KT-81938 Fixed