[FIR] Set proper type arguments for qualified accesses in completion
The UBV checker fails to detect the bound violation, because when it
reads `typeArguments` of the constructor call, it sees `AlphaKey` as
the second argument, not `BetaKey` despite it's written explicitly.
This happens because `typeArguments` are replaced inside
prepareQualifiedTransform` with substitutions from the new fresh type
variables via the `finalSubstitutor`.

^KT-80588