[FIR] Report the proper version of `UPPER_BOUND_VIOLATED` on typealiases
Even though it may seem that the reporting has become inferior, it's
worth saying that in the examples were it "worked properly", it did so
by accident (because the typealiases there trivially delegate the
parameters to the expansion). In more complicated scenarios the task
of determining the affecting original arguments is a bit more complicated
and if we do it wrong, we rist confusing the user.

^KT-80588