[K/N] Support generating calls to `@CGlobalAccess` in the compiler
See KT-79742 and the previous commit for the context.

This commit makes the compiler able to generate calls to property
accessors marked with `@CGlobalAccess`.

It generally does the same as 45cc5fe2 but now for C globals, which
are marked by cinterop with `@CGlobalAccess`.

In particular, the behaviour is configured by the same binary option

    cCallMode=indirect|indirectOrDirect|directOrIndirect|direct

Based on the value, the compiler decides whether it supports `@CCall`
and `@CGlobalAccess` on property accessors and which one it should
prefer if a function is marked with both.

The default value is `indirectOrDirect`, i.e. use `@CCall` unless only
`@CGlobalAccess` is available (which can happen if a cinterop klib is
generated with `-Xccall-mode direct`).

This commit also properly handles `@CGlobalAccess.Pointer` in the
compiler.

^KT-79742 Fixed