[FIR] Report INSTANCE_ACCESS_BEFORE_SUPER_CALL instead of UNRESOLVED_REFERENCE in class header
When trying to access a member declared inside a class in its header
(in super constructor call or delegation), we previously reported
UNRESOLVED_REFERENCE which was rather unhelpful.
Now, we report INSTANCE_ACCESS_BEFORE_SUPER_CALL and in the IDE, the
call will be resolved with error.
We achieve this by inserting an InaccessibleImplicitReceiverValue
in the tower context before (meaning further away than) the real
tower elements.
We rely on the trick with isContextCollectorMode to disable this
behavior in context collector mode, so that completion and reference
shortening are not affected.

#KT-15152 Fixed