[FIR] Fix rendering of inner types with generic outer classes
Inner types with generic outer classes are represented in FIR as having
all type arguments on the most inner class in a specific order
Foo<A, B>.Bar<C, D>.Baz<E, F> => Baz<E, F, C, D, A, B>.

Because cone types don't carry information is a type is inner and
whether type arguments belong to outer classes, we resolve the
types to their class-like symbols to reconstruct this information.

As a side effect, unresolved types are printed as fully qualified.

#KT-80741 Fixed