[K/N] Don't add `WrapperGenerationInfo` in cinterop when not needed
When generating StubIr for C globals, cinterop stores additional
information used by the C wrapper generator (`WrapperGenerationInfo`)
in a separate map.

In particular, the wrapper information is added unconditionally, because
wrappers were always generated.

But 79bc1be introduced a new mode (`-Xccall-mode direct`) where no
wrappers are generated at all. As a result, the global accessors can be
generated without the `@CCall` annotation but still added to the wrapper
generation maps.
The latter is incorrect, because `StubIrBridgeBuilder` expects
everything within those maps to require wrappers and have `@CCall`.

As a result, the C wrapper generation for such globals fails with an
exception, which luckily ignored.

This commit fixes this problem by adding `WrapperGenerationInfo` only
when necessary, i.e. when an accessor has `@CCall` (= requires a
wrapper).

^KT-79751