[FIR] Fix KT-71854
Drop a separate axis for `READY_FOR_FIXATION`. Originally it was
introduced just in case, because of a smelly `if` branching - a direct
artifact of the way the previous implementation had been written. Now
that it was dropped in the previous commit, we can also drop this one,
and move the "ready for fixation" readinesses to the proper place -
after all the other components that actually constitute the fixation
"readiness".

^KT-71854 Fixed
^KT-77939