[test] Migrate JS&Wasm invalidation tests to phased compilation pipeline
The changes in the `privateDeclarationLeakThroughDefaultParam` test
are driven by the fact that previously we forgot to check the diagnostic
reporter for errors. Now that we don't forget that,
the `IR_PRIVATE_TYPE_USED_IN_NON_PRIVATE_INLINE_FUNCTION` diagnostic
started being emitted on step 8.
To continue testing the IC behavior for this case, we had to disable
the corresponding language feature.

^KT-82204 Fixed
^KT-80696 Submitted