[Lowerings] Cleaner support of setters in UpgradeCallableReferences
In before handling of setters in generation of
IrRichReference for mutable property was
significantly relying on
1. Getter and setter have corresponding arugmnets on same positions
2. A prefix of arguments being bound.

While this is true in current language, it would not be true if we
support references for properties with contexts. It also is currently
not true if one of getter and setter is JvmStatic and other is not.

This refactoring simplifies this dependency, also removing dependency
between implementation details of upgrade lowering and
property reference lowering

^KT-82130 Fixed